Translator now optionally accepts parameter "supportedLangs". It can be a single value (e. g. "cs") or array of values (e. g. "[cs,en]").
In case user tries to access language which is not supported the translator will use default language instead.

This allows to use keywords (e.g. "users.manage") instead of default language (e. g. "Manage users") which should be
more consistent in time and allow to create a structure and keep context in localization files.